added state to timesheet.

+++ doneillspa/ClientApp/src/app/timesheets/timesheets.component.ts
@@ -59,7 +59,7 @@ export class TimesheetComponent {
     startOfWeek.setDate(this.currentDate.getDate() - (this.currentDate.getDay() - 1))
 
     //Setting up default timesheet and timesheet entries
-    this.activeTimeSheet = new Timesheet(0, localStorage.getItem('username'), localStorage.getItem('client_id'), localStorage.getItem('role'), startOfWeek);
+    this.activeTimeSheet = new Timesheet(0, localStorage.getItem('username'), localStorage.getItem('client_id'), localStorage.getItem('role'), startOfWeek,'New');
     this.newEntry = new TimesheetEntry("", "", "", "", "");
 
     this._projectService.getProjects().subscribe(result => {
@@ -116,6 +116,7 @@ export class TimesheetComponent {
       if (ts.owner == localStorage.getItem("client_id")) {
         this.activeTimeSheet.weekStarting = ts.weekStarting;
         this.activeTimeSheet.id = ts.id;
+        this.activeTimeSheet.status = ts.status;
         this.timesheetExists = true;
 
         for (let item of ts.timesheetEntries) {
@@ -182,28 +183,31 @@ export class TimesheetComponent {
   }
 
   removeTimesheetEntry(ts) {
-    this._timesheetService.deleteTimesheetEntry(ts).subscribe(
-      res => {
-        console.log(res);
-        if (this.selectedDay == "Mon") {
-          this.removeFromArrayList(this.monEntries, ts);
-        }
-        else if (this.selectedDay == "Tue") {
-          this.removeFromArrayList(this.tueEntries, ts);
-        }
-        else if (this.selectedDay == "Wed") {
-          this.removeFromArrayList(this.wedEntries, ts);
-        }
-        else if (this.selectedDay == "Thurs") {
-          this.removeFromArrayList(this.thursEntries, ts);
-        }
-        else if (this.selectedDay == "Fri") {
-          this.removeFromArrayList(this.friEntries, ts);
-        }
-        else {
-          this.removeFromArrayList(this.satEntries, ts);
-        }
-      }, error => this.errors = error);
+    //We can only delete timesheet entries while the timesheet is in a status of new.
+    if (this.activeTimeSheet.status == 'New') {
+      this._timesheetService.deleteTimesheetEntry(ts).subscribe(
+        res => {
+          console.log(res);
+          if (this.selectedDay == "Mon") {
+            this.removeFromArrayList(this.monEntries, ts);
+          }
+          else if (this.selectedDay == "Tue") {
+            this.removeFromArrayList(this.tueEntries, ts);
+          }
+          else if (this.selectedDay == "Wed") {
+            this.removeFromArrayList(this.wedEntries, ts);
+          }
+          else if (this.selectedDay == "Thurs") {
+            this.removeFromArrayList(this.thursEntries, ts);
+          }
+          else if (this.selectedDay == "Fri") {
+            this.removeFromArrayList(this.friEntries, ts);
+          }
+          else {
+            this.removeFromArrayList(this.satEntries, ts);
+          }
+        }, error => this.errors = error);
+    }
   }
 
   removeFromArrayList(array: TimesheetEntry[], ts: TimesheetEntry) {
@@ -359,6 +363,15 @@ export class TimesheetComponent {
         this.activeTimeSheet.id = res as number;
         this.timesheets.push(this.activeTimeSheet);
         $("#myNewTimesheetModal").modal('hide');
+      },
+      error => this.errors = error);
+  }
+
+  submitTimesheet() {
+    this.activeTimeSheet.status = "Submitted";
+    this._timesheetService.updateTimesheet(this.activeTimeSheet).subscribe(
+      res => {
+        console.log(res);
       },
       error => this.errors = error);
   }
